+++ frontend/src/metabase/query_builder/components/view/View/QueryViewNotebook/index.js
@@ -1,0 +1,1 @@
+export { default } from "./QueryViewNotebook";
